Move FixedOrientationOrthoViewController
